Allow knplabs/doctrine-behaviors 2
