+++ composer.json
@@ -36,14 +36,14 @@
         "doctrine/persistence": "<1.3.4",
         "doctrine/phpcr-odm": ">=3.0",
         "gedmo/doctrine-extensions": "<2.4.36",
-        "knplabs/doctrine-behaviors": "<1.0 || >=2.0",
+        "knplabs/doctrine-behaviors": "<1.0 || >=3.0",
         "stof/doctrine-extensions-bundle": "<1.1 || >=2.0"
     },
     "require-dev": {
         "doctrine/orm": "^2.5",
         "doctrine/phpcr-odm": "^1.5",
         "jackalope/jackalope-doctrine-dbal": "^1.5",
-        "knplabs/doctrine-behaviors": "^1.4",
+        "knplabs/doctrine-behaviors": "^1.4 || ^2.0",
         "matthiasnoback/symfony-config-test": "^4.2",
         "matthiasnoback/symfony-dependency-injection-test": "^4.0",
         "phpstan/extension-installer": "^1.0",
