+++ phpstan.neon.dist
@@ -10,3 +10,7 @@ parameters:
 
     excludes_analyse:
         - src/Test/DoctrineOrmTestCase.php
+        # Remove these lines when dropping support for KnpLabs/DoctrineBehaviors < 2.0
+        - tests/Fixtures/Model/Knplabs/TranslatableEntity.php
+        - tests/Fixtures/Model/Knplabs/TranslatableEntityTranslation.php
+        - tests/Fixtures/Traits/ORM/Knplabs/ArticleTranslation.php
